libopae++: Fix fpga_resource::enumerate_tokens
Change logic so that if tests if option_map list is not empty and then
it tries to convert them to filters (fpga_properties). After
iterating over the list of option_map objects, check the resulting list
of filters and if it is empty, then add one filter of the object type.