libopae: Ignore unreadable afu_id while enumerating
During enumeration, unsuccesfully trying to read an AFUID from the
'afu_id' sysfs node wil no longer fail the entire enumeration; instead,
it will simply not add the corresponding FPGA_ACCELERATOR to the list of
enumerated devices. This avoids failures when trying to enumerate
anything while an AFU is unresponsive (i.e. in reset, or unavailable due
to a port error).