fpgadiag: Move setting of CSRs after cache priming
Move the calls to set the CSRs of the actual NLB3 test after doing any
cache priming (cooling/warming).
Priming the cache will run NLB3 tests with a slightly different set of
CSR values. By moving the setting of the test CSRs after this, we can
ensure that the correct values are read used by the actual test.

* fpgadiag: Remove "#include <cassert>" in nlb3.cpp