hssi: remove unecessary przone writes in load cmd
Two przone writes were previously added to the GOTO command sequence as
a workaround to reset the accelerator. These were also being logged in
the mmio stream when generating a c header for BIOS integration.
Because these additional writes don't call hssi_ack and because hssi_ack
is the only place where a comma+newline is emitted to the header log,
the result is a header with some syntax errors.  This change removes
those unecessary writes and fixes the c header that is generated.