Add object ID property
The object ID is a system-wide unique 64 bit integer that uniquely
identifies a resource. In concept, it is similar to a token, but as a 64
bit value it is easier to pass between processes (or on a tool command
line).

The implementation of the object ID (i.e. how it is generated) is up to
the library - this commit uses a combination of the device's
major/minor number.

Note that applications should not make any assumptions about how the
object ID is generated, only that it is unique.