mux.py : Change mux.py to allow single configs
Change logic in mux.py that will generate a config file compatible with
sw components nlb0, nlb3, nlb7, etc. for running standalone.
Also changes the default command line switch for specifying config
( from -m to -c ) as this is the default switch for sw components.

In particular, it now generates options dictionaries dynamically based
on the args.engines
If this list of dictionaries is just one, it will include just the
dictionary in the final configuration - leaving out fpgamux specific
meta data.

This also includes:
* A fix for validating nlb3 configurations
* A fix to call process object communicate (remove wait for timeout)
  * This was necessary for cases that generate a lot of stdout
    That was locking up the proces
  * This means processes may run forever - will put in correct timeout
    fix later