c++utils: Add file logging to log
Add ability (to logger) to log to files

This includes:

* Adding a second constructor that takes in a file name
* Adding an open function that takes in a file name

Both the second constructor and the open function will use the
filename to open the filestream member variable. This will get
used as the sink if it is not null.

This also adds a flush function to the logger that propagates to the sink