libopae++: Add mmio_resource class
Add mmio_resource class that bypasses opae driver calls to perform mmio
operations and instead opens the mmio resource and uses this file
descriptor to map the mmio space.

This same technique is used by hssi fme class but this is not specific
to fme. It can be used to open the port device as well.
The fme class in hssi should probably be replaced with
this class - the only thing that the static fme::open function does is
check the socket id in the capability csr