libopae++: Fix hang issue in fpga_event::poll
* libopae++: Fix fpga_event::poll to not hang

Change the call to epoll_wait to use a timeout of 0.
This causes the call to return immediately, giving the caller the
opportunity to check for events at a finer resolution than the minimum
epoll timeout of 1 msec.

This also allows the poll loop to terminate when cancel is requested
from another thread.

* libopae++: Update fpga_event::poll

Update timedout variable to be true when
- elapsed time > timeout
- no events triggers