fpgainfo: Major refactor of fpgainfo
This change makes fpgainfo features to be more OO.
To do this, the sysfs module has been completed to wrap sysfs directory as
classes that can read/write to a syfs nodes within.

This also adds json serialization to be used for automation purposes or
integration with other tools.
The biggest change is representing syfs "files" as Python properties.

All the `sysfs_resource` subclasses have properties for "files" within the
directory.

For the case of fme_errors and port_errors classes, those are more
dynamic and will check the existence of the sysfs "file" before making
it an object property.

Finally, this adds ability to filter resource by BDF.

* Update fpgaerr.py

Fix typo - change `prot` to `port`

* fpgainfo: Fix typo in sysfs

This is for the case when the error files don't exist for the reported
revision. This adds a `*` to expand the errfiles item because the
argument is a `*` arg

* Fix pylint issues in fpgainfo