fpgabist: Add tool for built-in self-test
Introducing Built-in Self-Test:
Tool for self diagnosis of discrete FPGA card

usage: fpgabist [-h] [-i DEVICE_ID] [-b BUS] [-d DEVICE] [-f FUNCTION]
                gbs_paths [gbs_paths ...]

Use a greenbit stream (gbs) to initiate a self-test mode for the BIST
tool.

Currently supports nlb3 and dma AFUs. These AFUs run an fpgadiag and
dma memory test, respectively. In addition, Fpga system information is
printed to the console including, the number of found Fpgas and bdfs,
the Port and FME information, if any Port or FME errors exists and the
physical temperature of the card. The tool will attempt to perform a
Partial Reconfiguration and will terminate if this key feature is
unsuccessful.