fpgadiag: correct perf counter wrap when it occurs
nlb_stats relies on operator-() overloads for fpga_cache_counters and
fpga_fabric_counters to give it correct results for display. This
improvement checks for the case that the hardware counter has wrapped
and gives a value accounting for the counter wrap, rather than an errant
value exhibiting integer overflow.