Improve support for platform database in AFU JSON
This change is backward compatible with existing JSON and platform build scripts.

AFUs may be configured to run on multiple platforms. It makes little sense to
specify the platform's interface-uuid in a static AFU JSON file.  The schema no
longer requires "interface-uuid", making it possible to write an abstract AFU
JSON file and fill in the UUID at packaging time.

Typically, a packaging script for a particular platform sets interface-uuid.
The afu.py script is updated to provide syntax for specifying a key's full path,
making it possible to set afu-image/interface-uuid. The script is also updated
for legacy support: when just "interface-uuid" is specified it is assumed to be
"afu-image/interface-uuid".

afu-image/magic-no is now set automatically if not already present and is no
longer required to be set in every AFU's JSON file.