Add afu_sim_setup: a wrapper script for instantiating and configuring an ASE simulation (#43)
The new script is installed in OPAE bin and is aware of both the installation and source tree locations. It copies a base ASE tree to a target directory and then invokes generate_ase_environment.py. The script only supports the new flow, in which sources are specified explicitly in a file that is parsed by rtl_src_config.

In addition, generate_ase_environment.py is updated to auto-detect the simulator when no --tool is specified explicitly. If VCS is found, it is chosen.