Add option to install safe string library (#229)
The safe string library is used by some projects that depend on
OPAE.  This patch adds an option for installing the safe string
library as a static library that can be used in conjunction
with building applications that depend on OPAE.

It is not intended for providing widespread distribution of
the safe string library. The version of safe string included
here should only be used by projects that depend on OPAE as well.