Install safe_string static library in opae-devel
Include the safe_string library as an install target and as part
of the opae-devel package.  The reason is that this package is
needed by some of the AFUs (DMA for now).  Users of OPAE will
expect to be able to build the AFUs on their own system and will
need a way to link against safe_string library.

In the future if there is an alternative way to distribute safe
string it may be removed from the opae-devel package.  Users should
not rely on this as a way to redistribute this library.  But we
will likely use this technique for initial release of OPAE and
associated AFUs.