Always enable O3 for safe_string library
We noticed very large slowdown when switching from memcpy
to memcpy_s in the DMA AFU code.  Tests with OpenCL
MMD indicated that a large part of the speed drop could
be mitigated by compiling with O3.  This commit sets the
safe_string library to be compiled with O3 even when
building the rest of OPAE in debug mode.  I don't expect
that we should be debugging the internals of this library
any more than we would debug system memcpy implementation.
So always compiling with optimization seems like a good idea.