c++: Add guid_t struct, implicit conversion ops (#64)
* c++: Add guid_t struct, implicit conversion ops

Add a guid_t struct to wrap the guid type.
This uses std::array to store the guid data put also keeps a reference
to a props_ object. It uses the props_ variable to call C API functions
to get/set the GUID property.
It also copies the data when getting and setting so that it doesn't get
stale.

Add unit tests for properties and integrate those in CMake so that one
may use 'make test'

* c++: add << operator to guid_t

* c++: Update sample for use of guid_t type

* Update pvalue implicit conversion operator

Update pvalue implicit conversion operator to call the getter fn on the copy variable.

* c++: simplify pvalue<T> class

Simplify pvalue templated class to use its internal copy_ member
variable when calling the getter method and when using the `<<` stream
operator.