add "ctor/dtor" function in log.c
Add a static function, `init_logging`, with the `constructor`
attribute. This will cause the function to be called when the library is
loaded for the first time, thereby initializing the log level and the
log file if specified in the environment variables.

Also add a statc function `deinit_logging` that will close the log file
is it has been opened