opaec++: Fix guid_t and pvalue mutators (#149)
Member variable is_set_ tracks whether the local cached copy of the
object is in sync with the last call to the underlying property get
function. Each of the non-const members in these class must be careful to
reset is_set_ to false before modifying the cached copy and then to true
only on successful completion.