change travis build to use matrix (#159)
Use the matrix keyword to define four different job configurations to run in parallel.
These jobs run in parallel with their own dependencies.
For example, the job that only does tests the C coding style doesn't require any apt packages so they are omitted from it resulting in a faster job.
This is different because each job is configured explicitly whereas the previous job configuration used was more for stages where each job was configured the same. In other words, the old way ran apt-get install and pip install for each job. This configuration only execute those commands for the jobs that include them.