mmlink: fix read_mmr optimization issue (#179)
The read_mmr() function is used to obtain a pointer to a
memory-mapped IO address (void *) given an offset from the MMIO base.

In the previous implementation, callers were forced to cast the obtained pointer to a volatile type,
or else the optimization phases of the compiler produced invalid
code. This was observed as a regression test failure when the default
build type was changed to -DCMAKE_BUILD_TYPE=RelWithDebInfo.

The change converts read_mmr to a template function so that
no casting to volatile is required at the call site.