ASE: Add missing implementation for fpgaTriggerUmsg()
fpgaTriggerUmsg() was defined in <opae/umsg.h>, but not implemented in
libopae-c-ase. This commit adds an implementation that uses
fpgaGetUmsgPtr() to trigger an UMsg (similar to the libopae-c
implementation. It also contains an update to the Doxygen of
fpgaGetUmsgPtr() which still said that the call wasn't supported on ASE.

This addresses issue #176.