c++utils: add `any_value` class (#241)
* c++utils: add `any_value` class

Add `any_value` class to replace use of boost::any.
This class was developed by looking at some online references to tips on
how to to implement the any functionality in boost.

This is a step in removing dependency on boost. If boost libraries are
needed in the future, they can be added then.