fpgainfo: fix clear errors function (#265)
* fpgainfo: fix clear errors function

Add other sysfs nodes to the list of entries to clear when clearing
errors for fme.
To do this, the errors_feature was refactored to use a list of two
element tuples for the error nodes to clear. The first entry in the
tuple is the node to read from, the second is the node to write to for
clearing the errors.

Both fme_errors and port_errors classes were updated to specify their
respective tuples to clear.