merge master into feature/c++core (#278)
* Avoid clobbering CMAKE_C_FLAGS for bist tool (#250)

Set CMAKE_C_FLAGS for bist tool to include ${CMAKE_C_FLAGS} to avoid overriding default flags

* docs: add RPM extraction to install guide

Update the install guide with details of how to manually extract the
driver source files from an RPM. Without Debian package support,
some users are forced to use this method on Debian-based systems.

* CI: Automate release-tagged documentation upload process (#249)

-  Add documentation building to Travis
-  Install newer version of pandoc which has new features being used
-  Push documentation to opae.github.io when a tag is detected

* Add fpgabist to get checked for coding style (#246)

Format fpgabist C sources to comply with coding style and ensure all future code will also be checked for compliance.

* OPAE: fix gcc 7.3.1 build warnings

* OPAE: fix gcc 7.3.1 build warnings

Fix a handful of warnings that were exposed by -Werror -Wall when using
gcc 7.3.1 in Fedora 27. Add safestring dependency to hello_events and
scrub snprintf.

* Sync test files' license information with existing src code (#251)

Sync test files' license information with existing src code

* Move pClk frequency setting to the clocks class. (#262)

It was configured in the wrong place (the CCI-P interface class).

* fpgainfo: fix clear errors function (#265)

* fpgainfo: fix clear errors function

Add other sysfs nodes to the list of entries to clear when clearing
errors for fme.
To do this, the errors_feature was refactored to use a list of two
element tuples for the error nodes to clear. The first entry in the
tuple is the node to read from, the second is the node to write to for
clearing the errors.

Both fme_errors and port_errors classes were updated to specify their
respective tuples to clear.

* libopae: free MMIO mappings on close

Parameterize wsid_cleanup() with an optional callback so that work can
be done on each mapping when the list is being destroyed. Add call to
destroy handle::mmio_root in fpgaClose().

* Enable gtOpenClose.cpp in gtapi

* Tools: RAS support for discrete FPAG devices. (#242)

Changes:
 -Add RAS tool to  support discrete FPAG devices

* Enabled ctest output on failure. (#273)

* test: add test for auto-release of MMIO

Test that fpgaClose() automatically unmaps any mapped MMIO space.

* fpgadiag: fix "freq" option for bw calc. (#274)

Set the `frequency_` member variable for nlb0, nlb3, and nlb7 to the
value in the `freq` option. Without this, the frequency used for
bandwitdth calculations was always 400 MHz (ignoring user input).