fpgainfo: Add first version of fpgainfo in C
This adds a new cmake target, fpgainfo-c, which is compiled from C source.
When this is ready to replace the Python fpgainfo, we can rename the target/directory to fpgainfo.

This first version contains a main which parses the first level of arguments which include
the arguments to locate an accelerator or device (BDF and socket-id).

These arguments are used to populate a query filter (`fpga_properties`) which will later be used
when calling `fpgaEnumerate`

This main function also defines an array of `struct command_handler` objects where each item in
this array defines:
* command name - `{errors, port, fme, power, temp}`
* filter function - function called to add additional filters based on the sub-arguments
* run function - the function that performs the subcommand

When main identifies the `struct command_hander` object that pertains to the subcommand, it will
* Call the flter function to possibly add more filtering criteria
* Call `fpgaEnumerate` twice
  * Once to get the number of tokens for the filter, then allocate enough memory for tokens
  * Again to get the actual tokens
* Call the `run` function in the command handler (passing in the tokens and sub-arguments).

Finally, this adds to tests for `fpgainfo fme` and `fpgainfo port` commands