Add argsfilter static library
Add a static library, argsfilter, that can be used to parse common
command line arguments (bus, device, function, socket-id).

The only function this contains is fpgaPropertiesFromArgs which
can modify an fpga_properties filter by setting properties
when their corresponding arguments are found in the command line
vector.
Note, this will modify the argc/argv variables passed in so those
options are "removed" from the command line vector.

This is built as a static library with PIC turned on so it can be
linked into shared libraries.