Initialize error list when creating global token list
This commit moves the logic for populating the error
list to the point where the global token list is created.
This way, the error list is populated during the first
enumeration of a token. Since the error list for a
particular object type does not change, we don't
need to update it even for newly reconfigured
AFUs.

Also, fpgaCloneToken now does a shallow copy of
the error list.