Add force option for fpgaReconfigureSlot
This changes behavior of fpgaReconfigureSlot to try to open
an accelerator resource on the given slot and bail out if it can't.
This is to mitigate situations when a slot is reconfigured while
the accelerator is in use by another process.

This behavior can be skipped if the FPGA_RECONF_FORCE is given.
This also adds the corresponding command line option to fpgaconf