Add force option for fpgaReconfigureSlot (#329)
* Add force option for fpgaReconfigureSlot

This changes behavior of fpgaReconfigureSlot to try to open
an accelerator resource on the given slot and bail out if it can't
find it or open it.
The case for when it is not found should apply to when the
accelerator in that slot is assigned to a VM.
If it can't open it exclusively then it is in use by a non-virtualized
process on the same host.

This is to mitigate situations when a slot is reconfigured while
the accelerator is in use by another process.

This behavior can be skipped if the FPGA_RECONF_FORCE is given.
This also adds the corresponding command line option to fpgaconf.
When the force flag is on then the previous behavior is kept and the
slot will be reconfigured regardless of it being in use.

Also, remove `flags` variable as UNUSED_PARAM

(cherry picked from commit 3386aede2a65ef733864c95a075d714f30f106db)
Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>