Compile Qsys when configuring ASE simulation. (#388)
This eliminates the last significant difference between the ASE configuration
provided in OPAE and the ASE configuration shipped with discrete cards.  It was
a source of confusion in documentation, since we generally advised the use of
afu_sim_setup, which invokes generate_ase_environment.py.  For AFUs with Qsys,
we suggested a different script.  All shipped AFUs can now use the same
afu_sim_setup for configuration.