Merge feature/c++core into master (#390)
* libopae-cxx-core: First commit of cxx-core library

* remove 2nd layer C++ related files from cmake

* c++core: rename read/write to read_csr/write_csr (#261)

* c++core: rename read/write to read_csr/write_csr

* Rename csr read/write functions in `handle` class  to be more explicit in the operation and specify that the functions are operating on CSRs. The new function names are:
  * read_csr32
  * write_csr32
  * read_csr64
  * write_csr64

* Update samples to use updated API functions.

* exclude c++ headers from codingstyle test

* remove higher-level C++ API includes from gtapi

* Add fpgad to be build and run during mock testing

* libopaecxx: remove csr_space from handle::open

This allows relying on the lazy mapping performed by the
fpgaReadMMIOxx() and fpgaWriteMMIOxx() accessors, enabling
access to any of the supported MMIO spaces associated with
the handle. Parameterize member fn mmio_ptr() with a csr_space
argument.

* opaecxx: add exception for all the result codes (#277)

* opaecxx: add exception for all the result codes
* Create an exception class for each result code (other than FPGA_OK) in
  OPAE C API that derives from except
* Create a template class that throws an exception of template argument type if fpga_result is not FPGA_OK
* Add a table of functions that maps each fpga_result code to an instance of the template function using the corresponding exception class as its template argument
* Add function `assert_fpga_ok` that calls the corresponding function from the table for a given result code
* Add a macro, `ASSERT_FPGA_OK` that calls assert_fpga_ok with the src location
* Add unit tests for using the macro for each exception class

* merge master into feature/c++core (#278)

* Avoid clobbering CMAKE_C_FLAGS for bist tool (#250)

Set CMAKE_C_FLAGS for bist tool to include ${CMAKE_C_FLAGS} to avoid overriding default flags

* docs: add RPM extraction to install guide

Update the install guide with details of how to manually extract the
driver source files from an RPM. Without Debian package support,
some users are forced to use this method on Debian-based systems.

* CI: Automate release-tagged documentation upload process (#249)

-  Add documentation building to Travis
-  Install newer version of pandoc which has new features being used
-  Push documentation to opae.github.io when a tag is detected

* Add fpgabist to get checked for coding style (#246)

Format fpgabist C sources to comply with coding style and ensure all future code will also be checked for compliance.

* OPAE: fix gcc 7.3.1 build warnings

* OPAE: fix gcc 7.3.1 build warnings

Fix a handful of warnings that were exposed by -Werror -Wall when using
gcc 7.3.1 in Fedora 27. Add safestring dependency to hello_events and
scrub snprintf.

* Sync test files' license information with existing src code (#251)

Sync test files' license information with existing src code

* Move pClk frequency setting to the clocks class. (#262)

It was configured in the wrong place (the CCI-P interface class).

* fpgainfo: fix clear errors function (#265)

* fpgainfo: fix clear errors function

Add other sysfs nodes to the list of entries to clear when clearing
errors for fme.
To do this, the errors_feature was refactored to use a list of two
element tuples for the error nodes to clear. The first entry in the
tuple is the node to read from, the second is the node to write to for
clearing the errors.

Both fme_errors and port_errors classes were updated to specify their
respective tuples to clear.

* libopae: free MMIO mappings on close

Parameterize wsid_cleanup() with an optional callback so that work can
be done on each mapping when the list is being destroyed. Add call to
destroy handle::mmio_root in fpgaClose().

* Enable gtOpenClose.cpp in gtapi

* Tools: RAS support for discrete FPAG devices. (#242)

Changes:
 -Add RAS tool to  support discrete FPAG devices

* Enabled ctest output on failure. (#273)

* test: add test for auto-release of MMIO

Test that fpgaClose() automatically unmaps any mapped MMIO space.

* fpgadiag: fix "freq" option for bw calc. (#274)

Set the `frequency_` member variable for nlb0, nlb3, and nlb7 to the
value in the `freq` option. Without this, the frequency used for
bandwitdth calculations was always 400 MHz (ignoring user input).

* Revert "merge master into feature/c++core (#278)" (#283)

This reverts commit 861ded226ff626cc39e83975c3031bad2f510a3c.

* fix merge error

* libopae++: add MMIO tests

Add tests for class handle's MMIO API.

* make use of opae c++ exceptions (#288)

* opaecxx: Throw exceptions in token.cpp

Add calls to `ASSERT_FPGA_OK` macro for fpga_results objects we get from
calling the OPAE C API

* opaecxx: update pvalue and properties to throw

Update pvalue and properties class implementations to use the
ASSERT_FPGA_OK macro to throw exceptions for fpga_result codes of not
FPGA_OK

* opaecxx: update events to throw

Update events class implementation to use the ASSERT_FPGA_OK macro which
will throw exceptions when fpga_result code is not FPGA_OK

* opaecxx: update dma_buffer to throw

Update dma_buffer class implementation to use the ASSERT_FPGA_OK macro
which will throw exceptions when fpga_result code is not FPGA_OK

* opaecxx: Fix gtest for properties

* Fix set_objtype test by setting the property in the properties object
* Change the get_model test to expect an exception of type not_supported
  as it is currently not implemented in libopae-c

TODO: Add more properties tests

* remove log.h|cpp from cxx-core library (#336)

It has been decided to not include logging APIs in the cxx-core library

* remove doxygen tags that reference mmio_region (#337)

* opaecxx: Add version API (#342)

* opaecxx: Add version API

This adds three static functions in a `version` class under `opae::fpga::types`
namespace.

* Add C++ version test file

* Apply clang-format on gtCxxVersion file

* Apply clang-format to c++ code (#374)

This PR touches a lot of files but is only about code formatting.
I ran clang-format on the relevant source directories to make this.

TODO: Add coding style check to CI

* c++core: Make handle::close a public function (#370)

* c++core: Make handle::close a public function

Also cleanup some doxygen markup comments
Update the documentation to include a note about the handle d'tor also calling `close`

* change copyright year to 2018 (#385)

* Remove simple_example.cpp (#392)

Seeing as there is little value in having this sample, this removes it from the c++ samples.

* c++core: Rename `dma_buffer` to `shared_buffer` (#393)

* c++core: Rename `dma_buffer` to `shared_buffer`

To avoid confusion with other DMA data structures, this renames the `dma_buffer` class to `shared_buffer`

* c++core: add install for libopae-cxx-core (#396)