c++core: clone token when creating a new handle
The `fpga_handle` data stucture makes a shallow copy of the token
used when calling fpgaOpen. This change makes a clone of the token
before calling `fpgaOpen` and saves it in the C++ handle object.
This will ensure the lifetime of the token structure used to open
will be as long as the handle itself.