Change test configuration for naming convention
Add logic in test configuration to look for "ALL" in the filter.
Tests with "ALL" in the name should work with the mock driver.
Also fix the filters used in the tests/CMakeLists.txt when defining tests.
Finally, rename the cxx properties tests to use Props in the feature section of the name