refactor enumeration code (#428)
This changes the enumeration logic to only read afu_id from a port if it
has to. The conditions to when it should read afu_id are:
* No filters are given
* At least one filter includes FPGA_ACCELERATOR as the type
* At least one filter does not specify FPGA_ACCELERATOR as the type

To do this, the function `enum_fme_afu` has been split up in two:
* enum_fme
* enum_afu

The function enum_afu only gets called when the conditions listed above
are met.