pyopae: first version with properties
.clang-format is defines the coding standard to be used with clang-format.

pyopae/CMakeLists.txt contains the cmake logic to call into the cmake
function `create_python_dist`. It is also responsible for setting up
the virtual Python environment using virtualenv.
It also tests compiling against pybind11 using the test_pybind_compile.cpp
file.

opae.cpp is the main file that uses pybind11 to define the Python module.
This first version defines the enumerations and the properties class along with
references to the properties accessor methods.
The properties accessor methods (and conversions) are found in the
pyproperties.* files.

opae/__init__.py is the top level package initialization routine and uses
pkgutil style namespace package format and is recommended for Python 2.3+ and
Python 3 support. Using this format, other distrubutions can be installed into
the opae namespace as long as those other disributions also include the file
opae/__init__.py with the same content.
See the following links for more information on namespace
packages.

https://packaging.python.org/guides/packaging-namespace-packages/
https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages

opae/api/__init__.py is the package initialization routine for the the opae.api package.
This is the namespace where the _opae.so Python extension will be placed and so the purpose
of the api/__init__.py routine is to bring in the types/functions in the _opae.so module.

pyproperties.h|cpp contain the interfaces and implementations for the functions
that implement the Python bindings defined in opae.cpp.

setup.py is the template used to build the Python extension. This is what is
processed by setuptools both for building the extension as well as creating the
distribution package.

test_pyopae.py contains unit tests for the Python bindings.