pyopae: Add handle class binding
Add pyhandle.h|cpp which contains the interfaces and implementations for the
docstrings as well as some of the wrapper code.
The only methods that need to be wrapped here are:
* reconfigure
* __bool__

The `reconfigure` wrapper uses the Python C API to convert the file object to a
FILE* object. Once converted, the GBS is read into a buffer object which is
then passed to the C++ `handle::reconfigure` function

The __bool__ wrapper gets the C type and checks if it is NULL pointer.
If the underlying C type is a NULL pointer then it means the handle must have
been closed (or is invalid).