pyopae: Add shared_buffer class binding
Add pyshared_buffer.h|cpp which contains the interfaces and implementations for the
docstrings as well as some of the wrapper code.
The only methods that need to be wrapped here are:
* memoryview
* buffer protocol

The `memoryview` wrapper uses the pybind11 memoryview class to convert the file object to a
Python memoryview object.

The buffer protocol implementation is done by calling the
pybind11::buffer_protocol() function when defining the class in opae.cpp.