pyopae implementation (#460)
* add cmake create_python_dist function

* pyopae: first version with properties

.clang-format is defines the coding standard to be used with clang-format.

pyopae/CMakeLists.txt contains the cmake logic to call into the cmake
function `create_python_dist`. It is also responsible for setting up
the virtual Python environment using virtualenv.
It also tests compiling against pybind11 using the test_pybind_compile.cpp
file.

opae.cpp is the main file that uses pybind11 to define the Python module.
This first version defines the enumerations and the properties class along with
references to the properties accessor methods.
The properties accessor methods (and conversions) are found in the
pyproperties.* files.

opae/__init__.py is the top level package initialization routine and uses
pkgutil style namespace package format and is recommended for Python 2.3+ and
Python 3 support. Using this format, other distrubutions can be installed into
the opae namespace as long as those other disributions also include the file
opae/__init__.py with the same content.
See the following links for more information on namespace
packages.

https://packaging.python.org/guides/packaging-namespace-packages/
https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages

opae/api/__init__.py is the package initialization routine for the the opae.api package.
This is the namespace where the _opae.so Python extension will be placed and so the purpose
of the api/__init__.py routine is to bring in the types/functions in the _opae.so module.

pyproperties.h|cpp contain the interfaces and implementations for the functions
that implement the Python bindings defined in opae.cpp.

setup.py is the template used to build the Python extension. This is what is
processed by setuptools both for building the extension as well as creating the
distribution package.

test_pyopae.py contains unit tests for the Python bindings.

* pyopae: add token class to python binding

Add properties.get static method overload which uses a token object to get
properties from.

* pyopae: Update doc strings to raw strings

This update is needed to improve the formatting when generating documentation.

* pyopae: Change static functions to init

Change static factory functions for properties to init functions.

* pyopae: Add handle class binding

Add pyhandle.h|cpp which contains the interfaces and implementations for the
docstrings as well as some of the wrapper code.
The only methods that need to be wrapped here are:
* reconfigure
* __bool__

The `reconfigure` wrapper uses the Python C API to convert the file object to a
FILE* object. Once converted, the GBS is read into a buffer object which is
then passed to the C++ `handle::reconfigure` function

The __bool__ wrapper gets the C type and checks if it is NULL pointer.
If the underlying C type is a NULL pointer then it means the handle must have
been closed (or is invalid).

* pyopae: Add shared_buffer class binding

Add pyshared_buffer.h|cpp which contains the interfaces and implementations for the
docstrings as well as some of the wrapper code.
The only methods that need to be wrapped here are:
* memoryview
* buffer protocol

The `memoryview` wrapper uses the pybind11 memoryview class to convert the file object to a
Python memoryview object.

The buffer protocol implementation is done by calling the
pybind11::buffer_protocol() function when defining the class in opae.cpp.

* pyopae: Add event binding class

* make pybind compile test a repo file

* address pycodingstyle issues

* Add __init__.py and setup.py to exclude patterns for pycodestyle and
pylint
* Remove extra command in pyopae/setup.py

* Update test_pybind_compile.cpp

Address unused parameter warning

* pyopae: Add sphinx doc page, improve docstrings

* Add a shinx rst page to hold the Python API documentation
* Improve the docstrings in the pyopae classes.

* pyopae: fix reconfigure - replace fgets with fread

fgets stops reading from the file stream when it encounters a newline char.
This replaces it with fread to read the entire GBS buffer.

* pyopae: rename `api` namespace to `fpga`

* pyopae: move static methos to module methods

* Change enumerate, open, register_event and allocate_buffer from being class methods
  and make them module level methods in the opae.fpga namespace package.
* Update opae.fpga namespace package to explicitly import and include types and
  methods
* Remove FPGA_ from enumerations as it seems reduntant now that they are part
  of the opae.fpga namespace

* pyopae: Add context manager support to handle

Implement the __enter__ and __exit__ functions in the handle class.
This gives it context manager support (using the `with` statement).

* pyopae: fix typos spotted in PR