pyopae: enumerate w/ kwargs
Add an overload method of enumerate that takes in kwargs for the filter.
This eliminates the need for creating properties objects before calling
enumerate and allows client code to call enumerate with kwargs only.

This change also creates pytoken.h|cpp files which contain the code for the
documentation and implementation of token methods.