pyopae: Change the default build system
Temporarily disable building with setup.py and instead build the OPAE Python
bindings directly with cmake. This is done to test building with different
flags, especially building with coverage turned on and getting test coverage
info.

Add the required Python version as a cmake variable. This can later be used for
building the Python bindings with different Python versions.

This assumes that pybind11 and nose2 Python modules are installed in order to
get include paths to pybind11 and also to execute tests.
The travis jobs have been updated to install these modules.

Add a ctest in tests/CMakeLists.txt to call nose2 with the tests.

Add a "finish" function in the coverage script and move the processing of
coverage data into this. Change the script to call finish when exiting so that
this is called regardless of errors during testing. This will allow getting
coverage data even on test failures.