pyopae: add Python3 support
Because the Python API is slightly different in Python 3, the logic for
converting a file object from Python to a FILE* needs to be updated to account
for the Python 3 API (which includes `PyObject_AsFileDescriptor` function).

Also, the tests need to be modified to account for some differences in Python
functions/types in Python 3 vs Python 2.7.