pyopae: Add context manager helper class (#518)
Add a singleton class, `buffer_registry` which is designed to be called by the
context manager enter function to register a handle. It is then called again by
the shared_buffer allocate function to associate the resulting buffer with the
handle. The context manager exit function should then call the deregister
function in the buffer_registry to release buffers associated with a handle.