mock: Add fpgamock.c (to test mock in an exe)
Add fpgamock.c and build it into an executable, fpgamock.
The purpose of this is simply to test that building mock.c into an
executable will result in the functions defined in mock.c being called.

This has been tested in the debugger and I was able to step into
functions in mock.c.