Add generic functions for reading/writing objects
Add six functions for reading and writing from/to an object in an FPGA
resource. The objects are identified by a string key.
For Linux, this is equivalent to reading/writing to sysfs
files with 'key' as a path relative to the resource directory in sysfs.

For each operation (read, write), three functions are defined.
* One for operating on a 32-bit value
* One for operating a 64-bit value
* One for operating on a sequence of bytes