Refactor RW Bytes operations, add call to `close`
Change ReadBytes operation to use pread (partial read).
The reason for this change is becuase these calls don't change the file
offset (useful in multi-threaded applications).

Change WriteBytes operation to read the contents of the sysfs file,
modify it with the bytes to write, then write the contents back to the
file. This is recommnded in the Linux kernel sysfs docs:
https://www.kernel.org/doc/Documentation/filesystems/sysfs.txt

_NOTE_: Calling pread with an offset of 0 will cause the kernel driver
to reload the buffer.