Rename sys file, implement RW Bytes ops (#593)
* Rename sys file, implement RW Bytes ops

* Rename sys.h and sys.c to sysobject.h and sysobject.c
* Implement fpgaWriteObjectBytes, and fpgaReadObjectBytes functions
* Also change fn signature of fpgaWriteObjectBytes to use len output
param
* Add two helper functions to perform interrupt-able IO
  These functions continue reading/writing if the call to read/write is interrupted (as indicated by -1 result and errno). When either functions is interrupted, it continues to retry its IO operation.
* Change API spec of `fpgaReadObjectBytes`
  Use null buffer output pointer to indicate request of byte size
  necessary

The ReadBytes operation uses pread (partial read).
The reason for this change is becuase these calls don't change the file
offset (useful in multi-threaded applications).

The WriteBytes operation will first read the contents of the sysfs file,
modify it with the bytes to write, then write the contents back to the
file. This is recommnded in the Linux kernel sysfs docs:
https://www.kernel.org/doc/Documentation/filesystems/sysfs.txt

_NOTE_: Calling pread with an offset of 0 will cause the kernel driver
to reload the buffer.