Add `instance` to internal _fpga_token (#604)
* Add `instance` to internal _fpga_token

Add an instance number field to the internal sturcture, _fpga_token.
Parse that from sysfs path during enumeration and store it in the token.
This can then be used in other calls that require formatting paths using
that number.

When parsing the instance number, we do a reverse search for the '.'
character. Advance ptr by one before calling strtoul to parse the
instance number.