Add test_system and new mock.c
testing/mock.c calls C functions declared in c_test_system.
Those functions are implemented in test_system.cpp and call into
test_system singleton object for calls like open and opendir.

Unit tests (in testing/unit directory) are to call the test_system to
initialize it. For example, munit-enum-c.cpp initializes the test system
and also prepares a mock sysfs directory (as a subdirectory of cwd).

Calls destined to sysfs paths that are normally handled by the kernel
driver are handled by test_system. Calls to open devices in /dev, are
stashed in a map, keyed off the file descriptor, and are used when ioctl
is called with them.