userclk: Accept high, low, or both frequencies
Previously, the userclk tool would only process the high
frequency parameter, ignoring any low frequency parameter.
The high frequency was passed to the set_userclock libary
function as both the high and low frequency parameters.

Now that support for the IOPLL user clock is added, this
scheme doesn't work as well, because whereas the fpll programming
was based on the high frequency, the iopll programming is based
on the low frequency.

This change processes the high and/or low frequency. If only one
is provided, the other is calculated. If both are provided, then
an approximate relationship of (high = 2 * low) is enforced.

Additionally, the high and low frequencies, as measured by the
hardware, are displayed upon completion. If no frequency is
specified on the command line, then the tool effectively does
a read of the current frequencies.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>