userclk: Implement 10MHz to 600MHz range for iopll
Change the table of configuration data for the IOPLL to implement
a range of 10MHz to 600MHz. Our method of limiting the 2x clock
will be to use the same frequency as the 1x clock when the 1x
frequency is above 300MHz.

Also remove a check that allows the high freq to be passed in as
zero.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>