Add token_list tests, malloc override
Implement malloc in test_system.cpp  to check for a flag for returning
null. If the flag is set to true, it will reset it to false and return
null. The expected use case is calling test_system::invalidate_malloc()
to set the flag to true before calling a function that calls malloc.
This allows for negative testing of code that handles malloc failures.