plugin-api: refactor properties API
Make the property get/set routines of the properties API local
functions of opae-c, because they don't need to be plugin-enabled.
Remove the notion of wrapped properties from opae-c. Adjust the
test cases accordingly.

After fpgaGetProperties[FromHandle] returns from the plugin call,
the parent token in the properties, if any, will be a raw token.
We need to wrap it before handing the properties back to the caller.