tests: augment the malloc intercept
Add an optional count and symbol name to test_system::invalidate_malloc().
This will cause malloc to fail after the Nth call when the return
address is that of the given string. Note that this will not work
for inlined functions, irrespective of whether inlining is disabled.
Add fpgaOpen() test case that uses the functionality.