plugin: fix bugs in fpgaGetProperties[FromHandle]
and fpgaUpdateProperties. These functions must call through the
adapter in order to clone the returned parent token before wrapping
it. Attempting to call fpgaPropertiesGetParent() was resulting in
a failed top-level fpgaCloneToken() call, because the input token
wasn't wrapped yet.
Fix a few more object leaks in the opae-c properties tests.