Add test_system::remove_sysfs function
Since test_system is extracting the mock sysfs structure in the fn
test_system::prepare_syfs, it should also remove the temporary dirctory
that it creates from that step. This change adds a function
`remove_sysfs` which simply removes the directory (with some error
checking). Because this function is called in the test_system::finalize fn,
this allows the duplicate code to be removed from all the tests.