plugin-api: track wrapped_token allocs/free them
When allocating an opae_wrapped_token to wrap a parent token resulting
from a call to fpgaGetProperties[FromHandle], we need to set a bit in
the properties object that reminds us to free it from
fpgaDestroyProperties, else the wrapper will be leaked.
In fpgaUpdateProperties, if we determine that we have allocated the
wrapper for a set parent token, then the wrapper can be reused.
In fpgaPropertiesSetParent, we do the same as in fpgaUpdateProperties,
however we cannot make any assumption about the new parent; so leave
that to the caller.