tests: Add test coverage for opae-c properties
Increase coverage for fpgaGetProperties[FromHandle].
Add intercept for calloc(), modeled after malloc() intercept.
When allocating an opae_wrapped_token to wrap a parent token resulting
from a call to fpgaGetProperties[FromHandle], we need to set a bit in
the properties object that reminds us to free it from
fpgaDestroyProperties, else the wrapper will be leaked.
In fpgaUpdateProperties, if we determine that we have allocated the
wrapper for a set parent token, then the wrapper can be reused.
In fpgaPropertiesSetParent, we do the same as in fpgaUpdateProperties,
however we cannot make any assumption about the new parent; so leave
that to the caller.
Fix wrapper leak in fpgaUpdateProperties. Add tests for corner cases.
fpgaEnumerate was calling fpgaPropertiesGetParent to obtain the parent
token from the input filters array. In doing so, the parent tokens were
being cloned and subsequently leaked. Fix that by accessing p->parent
directly when unwrapping the parent token.