Update fpgaflash.md (#467)
* Update fpgaflash.md

* added revision history comment

* Update fpgaflash.md

* Update fpgaflash.md

* added w bmc flash types

* Update fpgaflash.md

* Update fpgaflash.md

* Update fpgaflash.md

* Update fpgaflash.md

* Update fpgaflash.md

* Update fpgaflash.md

* 4 updates suggested by Matthew Gerlach in tech review

* Update fpgaflash.md

* ran spellcheck in Oxyten

* ran spellcheck in Oxyten

* Use a hash to track the wsid space. (#724)

Unique workspace IDs are assigned to each allocated shared buffer. The IDs are used
for discovering the PA of a buffer and for passing the buffer size to the kernel when a page
is deallocated.  The data structure was a list, which didn't scale.  A million allocated
pages took over 1.5 hours to free.  The new code accomplishes the same in about 4 seconds.