Support multiple cards (#468)
* Wrong index for sensor name

* Change default wait time for a card to appear.

* Fixed two issues:

1. Disallow multiple instances of pacd to be started on the same
machine.
2. Move sensor contexts so they are thread-specific.

* Fix thread interactions.

* Fix fscanf warning.

* Use a hash to track the wsid space. (#724)

Unique workspace IDs are assigned to each allocated shared buffer. The IDs are used
for discovering the PA of a buffer and for passing the buffer size to the kernel when a page
is deallocated.  The data structure was a list, which didn't scale.  A million allocated
pages took over 1.5 hours to free.  The new code accomplishes the same in about 4 seconds.

* Addressing review comments