Remove handle parameter from fpgaObjectGetObject
This change simplifies the fpgaObjectGetObject API by eliminating the
optional handle parameter and instead changing the behavior so that any
object created from a handle will inherit the handle.

The object api sample is updated to not use the handle and may be
updated later to allow freezing of objects.